refact : 폴더 구조 리팩토링 및 테스트 코드 주석처리
